feat: move handleSearchQuery to HOC App
